service/src/requirements.txt
Flask==2.3.2
openai==0.27.6
pinecone_client==2.2.1
protobuf==3.12.4
